fix: use nullish coalescing to fix priority 0 bug display
The monitory-ui js rendering has logic for displaying issue priority
with a fallback to 2 if there is no priority. That logic relies on an
`||` which will treat a falsey value like 0 as false. The result is an
issue with a priority of 0 will render on the ui as the fallback of P2.

This PR also adds P0 as a filter to the priority selector so P0s can
also be filtered on without requiring removing all the priority filters.