docs: fix incorrect docs for how to install Claude Code plugin from source
The Claude Code command `/plugin marketplace add .` doesn't work.  For
some reason Claude accepts `./beads` as an argument but not `.`, so
change it to that.

Also clarify the difference between shell commands and Claude Code
commands, and mention that if you want to install the plugin locally
in your code repo, you'll need to `cd` there first.